Separate ActiveStorage facade from Cellar machinery
